requirements.txt
arrow>=0.17.0
boto3>=1.16.13
Jinja2>=2.11.2
PyYAML>=5.3.1
requests>=2.24.0
cryptography>=3.2.1
